requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.1
pyyaml>=6.0
